+++ packages/agent/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Compaction summary prompts now include an Incomplete Todos section so pending/in_progress items from additional-context survive the cut ([#8874](https://github.com/can1357/oh-my-pi/issues/8874)).
+
 ## [18.3.1] - 2026-09-25
 
 ### Added